Update paper.md
Couple of corrections -- typos. Removed final trailing sentence: "This decoupling and publishing an `npm` package and an R package separately, we believe, is the appropriate way of working with modern".  Not sure if we want to say anything more / elaborate on the usefulness of decoupling JS (application) from R (back-end).